Add a host-side test harness, and fix the memory errors it found
The port had never been run. distingNT/test builds each algorithm into its own
executable with a stand-in for the firmware and the address and
undefined-behaviour sanitizers turned on. The harness plays the part of the
module: it asks the plug-in for its factories, allocates exactly the memory each
declares with guard regions around it, feeds parameter defaults and busses, and
renders three passes - the defaults, every input saturated, and rounds of
pseudo-random parameters. It checks the parameter tables against their pages,
that nothing writes outside its allocation, that every sample is finite, and
that an instrument makes a sound.

All sixteen algorithms pass, over ten random seeds. Getting there meant fixing
memory errors in the shared emulators, all of which affect the Rack build too:

- BLIPBuffer::read_sample() moved one element too many and cleared one element
  past the end of its buffer, on every sample of every buffer. In the Rack build
  that write lands in the next buffer's first field; on the NT it landed in the
  chip emulator and zeroed an oscillator's output pointer.
- BLIPSynthesizer::adjust_impulse() ran its phase loop one step wide at both
  ends, reading before and writing past the impulse table, and never reached the
  phase-0.5 case that halves the error correction.
- Ricoh2A03::Oscillator::reset() was missing its zero assignment, copying the
  uninitialised fourth register over the other three and leaving the
  reg_written flags holding whatever was in memory.
- NintendoGBS sized its register array from the difference of its address
  bounds rather than the span, one short of the inclusive range write() accepts.
- DigitalOscillator took its wave-table note straight from an unbounded pitch,
  so a zero frequency or a large CV indexed past the table and segfaulted.
- The S-DSP, the echo and the YM2612 shifted signed values left and overflowed
  signed products. The shifts became multiplies by the same powers of two,
  exact at these magnitudes, and the products are widened before clamping.

Super VCA's loudness compensation and Super Sampler's unmasked nibble packing
both look wrong but are left alone, since changing them changes what the modules
sound like; both are written up in distingNT/README.md.

+++ distingNT/plugins/SuperSampler.cpp
@@ -190,11 +190,17 @@ static void setupSourceDirectory(uint8_t* ram) {
                 i < 2 * SonyS_DSP::BitRateReductionBlock::NUM_SAMPLES;
                 i += 2) {
             // hyaw_sample() reads zero past the end of the sample, which the
-            // last block runs into
-            const uint8_t hi = SonyS_DSP::hyaw_sample(i + 16 * block_index);
-            const uint8_t lo =
-                SonyS_DSP::hyaw_sample(i + 16 * block_index + 1) << 4;
-            block->samples[i / 2] = hi | lo;
+            // last block runs into. The samples are signed, so the packing is
+            // done unsigned: shifting a negative value left is undefined, and
+            // the sign extension of a negative sample is what sets the high
+            // nibble here, as it does in the Rack build.
+            const unsigned hi = static_cast<unsigned>(
+                SonyS_DSP::hyaw_sample(i + 16 * block_index)
+            );
+            const unsigned lo = static_cast<unsigned>(
+                SonyS_DSP::hyaw_sample(i + 16 * block_index + 1)
+            ) << 4;
+            block->samples[i / 2] = static_cast<uint8_t>(hi | lo);
         }
     }
 }
